232 Qwerty Key input

 - Show qwerty keys on top of virtual piano keys.
 - These qwerty keys generate music KB events, just as for the virtual piano keys and (TODO) midi input.
 - We now sound/silence notes in the game state where we consume the music KB event, so we can adjust the octave of the pitch.
 - All key events are handled by a central KeyInputHandler which manages mappings and handles auto-repeat.

+++ Source/Animals/AnimalController.cpp
@@ -183,7 +183,7 @@ std::vector<RCPtr<PFNpc>> AnimalController::AddPetsForGameRound(
 
     // TODO srand on game round ID to make this the same every time for 
     //  any given game round
-    int r = Amju::RandomInt(PET_TYPES.size());
+    int r = Amju::RandomInt(static_cast<int>(PET_TYPES.size()));
 
     auto petType = PET_TYPES[r];
     auto pet = dynamic_cast<Pet*>(AddAnimalFixedZ(petType, z));
